+++ com.unity.render-pipelines.universal/Runtime/DeferredLights.cs
@@ -285,7 +285,7 @@ namespace UnityEngine.Rendering.Universal.Internal
             );
 
             {
-                CommandBuffer cmd = CommandBufferPool.Get();
+                var cmd = renderingData.commandBuffer;
                 using (new ProfilingScope(cmd, m_ProfilingSetupLightConstants))
                 {
                     // Shared uniform constants for all lights.
@@ -315,7 +315,7 @@ namespace UnityEngine.Rendering.Universal.Internal
                 }
 
                 context.ExecuteCommandBuffer(cmd);
-                CommandBufferPool.Release(cmd);
+                cmd.Clear();
             }
 
             Profiler.EndSample();
@@ -528,7 +528,7 @@ namespace UnityEngine.Rendering.Universal.Internal
             if (m_StencilDeferredPasses[0] < 0)
                 InitStencilDeferredMaterial();
 
-            CommandBuffer cmd = CommandBufferPool.Get();
+            var cmd = renderingData.commandBuffer;
             using (new ProfilingScope(cmd, m_ProfilingDeferredPass))
             {
                 // This does 2 things:
@@ -555,9 +555,6 @@ namespace UnityEngine.Rendering.Universal.Internal
             CoreUtils.SetKeyword(cmd, ShaderKeywordStrings.AdditionalLightShadows, renderingData.shadowData.isKeywordAdditionalLightShadowsEnabled);
             CoreUtils.SetKeyword(cmd, ShaderKeywordStrings.SoftShadows, renderingData.shadowData.isKeywordSoftShadowsEnabled);
             CoreUtils.SetKeyword(cmd, ShaderKeywordStrings.LightCookies, m_LightCookieManager.IsKeywordLightCookieEnabled);
-
-            context.ExecuteCommandBuffer(cmd);
-            CommandBufferPool.Release(cmd);
         }
 
         // adapted from ForwardLights.SetupShaderLightConstants